Updates:
- Setting the gender of a Sim will now also set the proper gender trait
on the Sim too within CommonGenderUtils
- CommonSimPregnancyUtils.can_impregnate and
CommonSimPregnancyUtils.can_be_impregnated now properly check the gender
of an animal Sim for whether they may impregnate or be impregnated.
- Added a class log to CommonTraitUtils (Each function was updated to
a classmethod from a static method to reflect this!)
- Removed the FOX species from being allowed to perform the social super
pet socials provided by S4CL, custom ones for Foxes will be provided in
the future.