Updates:
- Fixed an error when converting a Location to a CommonLocation where
the parent_ref did not have a provided_routing_surface property
- Fixed an issue where the Cheat menu would be visible without Cheats
being enabled when holding the SHIFT key.
- Added a version number to logged files (Assuming the mod specifies a
version number)
- Fixed an issue with interactions not having proper visible names when
printing them in the notification
- Fixed an issue when cancelling all running interactions and the Sim
adds something to their queue during that.
- Added more details to Test Result verbose log messages in interactions to
indicate which level of interaction is being called and where the result
of the Test Result is returning from (Which class in particular).
- Added MISSING_IMAGE_ICON to CommonIconId